% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eformat.data.frame}
\alias{eformat.data.frame}
\title{Guess desired format for data frame}
\usage{
\method{eformat}{data.frame}(object, colnames = NULL)
}
\arguments{
\item{object}{data frame to format}

\item{colnames}{alternative column names}
}
\description{
Guess desired format for data frame
}

