% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eformat}
\alias{eformat}
\title{Default formats for the exhibit function}
\usage{
eformat(object, ...)
}
\arguments{
\item{object}{output of exhibit function to format}

\item{...}{other options}
}
\description{
\code{eformat} defines certain default ways to format
objects created by functions in the exhibit function.
}

