% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.ata}
\alias{exhibit.ata}
\title{Returns an age to age development triangle matrix}
\usage{
\method{exhibit}{ata}(object, selection = NULL)
}
\arguments{
\item{object}{object of class ata from the \code{ChainLadder} package}

\item{selection}{Selected development factors.  Should be supplied
as a numeric vector of the same length as the number of development periods
in the ata development triangle or the number of development periods + 1 
(the + 1 will be used as the tail factor).}
}
\description{
Returns an age to age development triangle matrix
}
\examples{
library(ChainLadder)
dev_tri <- ata(RAA)

# No selected development factors
exhibit(dev_tri)

# with selected development factors
exhibit(dev_tri, selection = c(3.0, 1.65, 1.30, 1.2, 1.12, 1.05, 1.035, 1.015, 1.01))

# with selected development factors includeing tail factor
exhibit(dev_tri, selection = c(3.0, 1.65, 1.30, 1.2, 1.12, 1.05, 1.035, 1.015, 1.01, 1.05))
}

