% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.BootChainLadder}
\alias{exhibit.BootChainLadder}
\title{Returns BootChainLadder summary information in a data frame}
\usage{
\method{exhibit}{BootChainLadder}(object)
}
\arguments{
\item{object}{object of class BootChainLadder generated from \code{ChainLadder} package}

\item{eformat}{logical; whether or not to use default exhibit format}
}
\description{
Returns BootChainLadder summary information in a data frame
}
\examples{
boot_object <- BootChainLadder(RAA)

exhibit(boot_object)
}

