% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.glmReserve}
\alias{exhibit.glmReserve}
\title{Returns glmReserve summary information as a data frame}
\usage{
\method{exhibit}{glmReserve}(object)
}
\arguments{
\item{object}{object of class glmReserve generated from \code{ChainLadder} package}
}
\description{
Returns glmReserve summary information as a data frame
}
\examples{
                   
glm_object <- glmReserve(GenIns)

exhibit(glm_object)
}

