% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.MackChainLadder}
\alias{exhibit.MackChainLadder}
\title{Returns MackChainLadder summary information in a data frame}
\usage{
\method{exhibit}{MackChainLadder}(object)
}
\arguments{
\item{object}{object of class MackChainLadder generated from \code{ChainLadder} package}
}
\description{
Returns MackChainLadder summary information in a data frame
}
\examples{
library(ChainLadder)
mack_object <- ChainLadder::MackChainLadder(RAA)

exhibit(mack_object)
}

