% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibit.R
\name{exhibit}
\alias{exhibit}
\title{Generic function for creating tabular data for report output}
\usage{
exhibit(object, ...)
}
\arguments{
\item{object}{object to turn into exhibit}

\item{...}{additional arguments}
}
\description{
\code{exhibit} defines certain default ways to arrange 
objects returned by \code{ChainLadder} and other
packages for presentation in a report.
}

