% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.triangle}
\alias{exhibit.triangle}
\title{Returns a cleaner development triangle for use in reports}
\usage{
\method{exhibit}{triangle}(object)
}
\arguments{
\item{object}{object of class triangle generated from \code{ChainLadder} package}

\item{eformat}{logical; whether or not to use default exhibit format}
}
\description{
Returns a cleaner development triangle for use in reports
}
\examples{
exhibit(RAA)
}

