% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAA}
\alias{RAA}
\title{RAA triangle provided with the \code{ChainLadder} package}
\format{An object of class \code{triangle} (inherits from \code{matrix}) with 10 rows and 10 columns.}
\usage{
RAA
}
\description{
RAA triangle provided with the \code{ChainLadder} package
}
\keyword{datasets}

