% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.ata}
\alias{exhibit.ata}
\title{Return a triangle for age to age development factors}
\usage{
\method{exhibit}{ata}(object, selection = NULL)
}
\arguments{
\item{selection}{optional selected development factors.  Should be supplied
as a numeric vector of the same length as the number of development periods
in the ata development triengle or the number of developments + 1 in the ata
development triengle, with the + 1 for the tail factor.  The tail factor can be
used to project future development for the most mature origin years.}

\item{ata}{object of class ata generated from \code{ChainLadder} package}
}
\description{
Return a triangle for age to age development factors
}
\examples{
# make sure to load the ChainLadder package with
# library(ChainLadder)
dev_tri <- ata(RAA)

# No selected development factors
exhibit(dev_tri)

# with selected development factors
exhibit(dev_tri, selection = c(3.0, 1.65, 1.30, 1.2, 1.12, 1.05, 1.035, 1.015, 1.01))

# with selected development factors includeing tail factor
exhibit(dev_tri, selection = c(3.0, 1.65, 1.30, 1.2, 1.12, 1.05, 1.035, 1.015, 1.01, 1.05))
}

