% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exhibit.ata}
\alias{exhibit.ata}
\title{Return a triangle for age to age development factors}
\usage{
\method{exhibit}{ata}(object, selection = NULL)
}
\arguments{
\item{selection}{optional selected development factors.  Should be supplied
as a numeric vector of the same length as the number of development periods
in the ata development triengle or the number of developments + 1 in the ata
development triengle, with the + 1 for the tail factor.  The tail factor can be
used to project future development for the most mature origin years.}

\item{ata}{object of class ata generated from \code{ChainLadder} package}
}
\description{
Return a triangle for age to age development factors
}
\examples{
tri <- as.triangle(ldf_data, origin = "origin",
                   dev = "dev", value = "paid_loss_only")
dev_tri <- ata(tri)
exhibit(dev_tri, selection = c(1.9, 1.2, 1.13, 1.075))

# with tail factor selected
exhibit(dev_tri, selection = c(1.9, 1.2, 1.13, 1.075, 1.1))
}

