% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exhibit.BootChainLadder}
\alias{exhibit.BootChainLadder}
\title{Returns BootChainLadder summary information in a data frame}
\usage{
\method{exhibit}{BootChainLadder}(object, format = TRUE)
}
\arguments{
\item{object}{object of class BootChainLadder generated from \code{ChainLadder} package}

\item{format}{whether or not to use default exhibit format}
}
\description{
Returns BootChainLadder summary information in a data frame
}
\examples{
tri <- as.triangle(recovery_ldf, origin = "origin",
                   dev = "dev", value = "paid_loss_only")

boot_object <- BootChainLadder(tri)

exhibit(boot_object)
exhibit(boot_object, format = FALSE)
}

