% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.BootChainLadder}
\alias{exhibit.BootChainLadder}
\title{Returns BootChainLadder summary information as a data frame}
\usage{
\method{exhibit}{BootChainLadder}(object)
}
\arguments{
\item{object}{object of class BootChainLadder generated from \code{ChainLadder} package}
}
\description{
Returns BootChainLadder summary information as a data frame
}
\examples{
library(ChainLadder)          
boot_object <- ChainLadder::BootChainLadder(RAA)

exhibit(boot_object)
}

