% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exhibit.glmReserve}
\alias{exhibit.glmReserve}
\title{Returns glmReserve summary information in a data frame}
\usage{
\method{exhibit}{glmReserve}(object, eformat = FALSE)
}
\arguments{
\item{object}{object of class glmReserve generated from \code{ChainLadder} package}

\item{eformat}{logical; whether or not to use default exhibit format}
}
\description{
Returns glmReserve summary information in a data frame
}
\examples{
tri <- as.triangle(recovery_ldf, origin = "origin",
                   dev = "dev", value = "paid_loss_only")

glm_object <- glmReserve(tri)

exhibit(glm_object)
exhibit(glm_object, format = FALSE)
}

