% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.glmReserve}
\alias{exhibit.glmReserve}
\title{Returns glmReserve summary information in a data frame}
\usage{
\method{exhibit}{glmReserve}(object)
}
\arguments{
\item{object}{object of class glmReserve generated from \code{ChainLadder} package}

\item{eformat}{logical; whether or not to use default exhibit format}
}
\description{
Returns glmReserve summary information in a data frame
}
\examples{
glm_object <- glmReserve(GenIns)

exhibit(glm_object)
}

