% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.MackChainLadder}
\alias{exhibit.MackChainLadder}
\title{Returns MackChainLadder summary information in a data frame}
\usage{
\method{exhibit}{MackChainLadder}(object)
}
\arguments{
\item{object}{object of class MackChainLadder generated from \code{ChainLadder} package}

\item{eformat}{whether or not to use default exhibit format}
}
\description{
Returns MackChainLadder summary information in a data frame
}
\examples{
tri <- as.triangle(ldf_data, origin = "origin",
                   dev = "dev", value = "paid_loss_only")

mack_object <- MackChainLadder(tri)

exhibit(mack_object)
}

