% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{exhibit}
\alias{exhibit}
\alias{exhibit-package}
\title{The exhibit package.}
\usage{
exhibit(odject, ...)
}
\arguments{
\item{...}{additional arguments}

\item{object}{object to turn into exhibit}
}
\description{
The exhibit package.

\code{exhibit} defines certain default ways to arrange
objects returned by the package \code{ChainLadder} package
and \code{lossdb} package functions for presentation in a report.
}

