% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exhibit.R
\docType{package}
\name{exhibit}
\alias{exhibit}
\alias{exhibit-package}
\title{The exhibit package.}
\usage{
exhibit(oject, format = TRUE, ...)
}
\arguments{
\item{format}{whether or not to use default `exhibit` format}

\item{...}{additional arguments}

\item{object}{object to turn into exhibit}
}
\description{
The exhibit package.

\code{exhibit} defines certain default ways to arrange
objects returned by the package \code{ChainLadder} package
and \code{lossdb} package functions for presentation in a report.
}
\details{
Many packages in R return list data structures that contain
a lot more information than is usually needed.  \code{exhibit} returns
a summary of this output as a data frame that can be easily displayed as
a table in a pdf or html document.
}

