% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibit.R
\name{exhibit}
\alias{exhibit}
\title{Generic similar to summary but always creates a data frame}
\usage{
exhibit(object, format = TRUE, ...)
}
\arguments{
\item{object}{object to turn into exhibit}

\item{format}{whether or not to use default `exhibit` format}

\item{...}{additional arguments}
}
\description{
\code{exhibit} defines certain default ways to arrange 
objects returned by the package \code{ChainLadder} package
and \code{lossdb} package functions for presentation in a report.
}

