% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainladder.R
\name{exhibit.triangle}
\alias{exhibit.triangle}
\title{Returns a development triangle as a matrix}
\usage{
\method{exhibit}{triangle}(object)
}
\arguments{
\item{object}{object of class triangle generated from \code{ChainLadder} package}
}
\description{
Returns a development triangle as a matrix
}
\examples{
                   
exhibit(RAA)
}

