% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exhibit.triangle}
\alias{exhibit.triangle}
\title{Returns a cleaner development triangle for use in reports}
\usage{
\method{exhibit}{triangle}(object)
}
\arguments{
\item{object}{object of class triangle generated from \code{ChainLadder} package}

\item{eformat}{logical; whether or not to use default exhibit format}
}
\description{
Returns a cleaner development triangle for use in reports
}
\examples{
tri <- as.triangle(ldf_data, origin = "origin",
                   dev = "dev", value = "paid_loss_only")

exhibit(tri)
}

